refactor(artist): 사용하지 않는 멀티타입 제거

+++ android/festago/presentation/src/main/java/com/festago/festago/presentation/ui/artistdetail/ArtistDetailViewModel.kt
@@ -37,7 +37,7 @@ class ArtistDetailViewModel @Inject constructor(
 
     private fun Stages.toUiState() = this.stage.map {
         StageUiState(
-            id = it.id.toLong(),
+            id = it.id,
             name = it.name,
             imageUrl = it.imageUrl,
             startDate = it.startDate,
